requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses==0.6
exceptiongroup==1.2.2
fastapi==0.112.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
pokedex.py==2.0.0
protobuf==5.27.3
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.28.2
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
